Build fix by updating connection properties
