+++ README.md
@@ -2,4 +2,3 @@
 
 [![Build Status](https://travis-ci.org/Seiferxx/wissen.svg?branch=v0.0.4_1812)](https://travis-ci.org/Seiferxx/wissen)
 
-Spring web application
