+++ pom.xml
@@ -25,10 +25,6 @@
   <dependencies>
     <dependency>
       <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-thymeleaf</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-data-mongodb</artifactId>
     </dependency>
     <dependency>
@@ -67,10 +63,10 @@
       <artifactId>spring-security-oauth2-autoconfigure</artifactId>
       <version>2.0.1.RELEASE</version>
     </dependency>
-    <!-- dependency>
+    <dependency>
       <groupId>de.flapdoodle.embed</groupId>
       <artifactId>de.flapdoodle.embed.mongo</artifactId>
-    </dependency> -->
+    </dependency> 
   </dependencies>
   
   <build>
