+++ src/main/java/com/seifernet/wissen/configuration/OAuth2ResourceServerConfiguration.java
@@ -24,7 +24,7 @@ public class OAuth2ResourceServerConfiguration extends ResourceServerConfigurerA
 	public void configure(HttpSecurity http) throws Exception {
 		http
 			.authorizeRequests()
-				.antMatchers("/api/tasks/*", "/api/tasks")
+				.antMatchers("/api/*", "/api/", "/api")
 					.fullyAuthenticated();
 	}
 }
