--- src/main/java/com/seifernet/wissen/util/ModelAttributes.java
@@ -1,14 +1,0 @@
-package com.seifernet.wissen.util;
-
-/**
- * Model attribute names
- * 
- * @author Seiferson (Cuauhtemoc Herrera)
- */
-public class ModelAttributes {
-	
-	public static final String CUSTOM_JS_SOURCE		= "customjsSource";
-	public static final String CUSTOM_JS_FRAGMENT	= "customjsFragment";
-	public static final String CONTENT_SOURCE		= "contentSource";
-	public static final String CONTENT_FRAGMENT		= "contentFragment";
-}
