+++ src/main/java/com/seifernet/wissen/util/URL.java
@@ -10,4 +10,5 @@ public class URL {
 	public static final String INDEX = "/";
 	public static final String EXPENSES = "/expenses";
 	public static final String DASHBOARD = "/dashboard";
+	public static final String NONE = "/none";
 }
