--- src/main/java/com/seifernet/wissen/util/WebResources.java
@@ -1,16 +1,0 @@
-package com.seifernet.wissen.util;
-
-/**
- * Web resources definitions
- * 
- * @author Seiferson (Cuauhtemoc Herrera)
- */
-public class WebResources {
-	
-	public static final String BASE_LAYOUT 	= "main_layout";
-	public static final String VOID			= "empty";
-	public static final String INDEX		= "index";
-	public static final String DASHBOARD	= "dashboard";
-	public static final String EXPENSES		= "expenses";
-	public static final String CUSTOM_JS 	= "custom_js";
-}
