+++ src/main/resources/application.properties
@@ -3,6 +3,7 @@ spring.data.rest.base-path = /api
 management.endpoints.web.exposure.include = *
 management.endpoint.shutdown.enabled = true
 server.error.whitelabel.enabled=false
+spring.data.rest.defaultMediaType = application/json
 
 info.app.name=Wissen
 info.app.version=0.0.4
@@ -37,6 +38,7 @@ wissen.profile = DEVENV
 
 logging.file = wissen.log
 logging.level.com.seifernet.wissen = INFO
+logging.level.com.seifernet.wissen.scheduler = WARN
 logging.level.org.springframework = WARN
 logging.level.org.mongodb = WARN
 logging.level.org.apache = WARN
