+++ src/main/resources/public/css/wissen.css
@@ -4,62 +4,6 @@ body > .ui.container {
 	margin-top: 3em; 
 }
 
-.floatxbutton{
-	position:fixed;
-	bottom:20px;
-	right:20px;
-}
-
-i.leaf.icon.greenicon {
-	color: #339933; 
-}
-
-span#latest.greenstatus {
-	color: #009933; 
-}
-
-span#latest.redstatus {
-	color: #cc3300; 
-}
-
-.hiddenf {
-	display: none !important; 
-}
-
-.tlistfelem {
-	padding-bottom: 7px !important;
-}
-
-.tlistelem {
-	padding-bottom: 7px !important; 
-	padding-top: 7px !important;
-}
-
-.tlistfcol{
-	margin-bottom:0px !important;
-}
-
-.tlistaddcol{
-	margin-bottom:0px !important;
-	margin-top:0px !important;
-}
-
-.tlistcol {
-	margin-bottom:0px !important; 
-	margin-top:0px !important
-}
-
-.floatfright{
-	float:right;
-}
-
-.unselectablef{
-  -moz-user-select: none;
-  -khtml-user-select: none;
-  -webkit-user-select: none;
-  user-select: none;
-}
-
 /* Semantic UI has these classes, however they're only applicable to*/
 /* grids, containers, rows and columns.*/
 /* plus, there isn't any `mobile hidden`, `X hidden` class.*/
